#243 - Share user's being when connection is active.

+++ src/Recollections.Api/Entries/Controllers/BeingController.cs
@@ -144,6 +144,9 @@ namespace Neptuo.Recollections.Entries.Controllers
         {
             string userId = User.FindUserId();
 
+            if (userId == entity.Id)
+                return BadRequest();
+
             await shareDeleter.DeleteBeingSharesAsync(id);
 
             db.Beings.Remove(entity);
